requirements.txt
asgiref
certifi
chardet
django-decouple
humanfriendly
idna
pyreadline
pytz
requests
sqlparse
typing-extensions
urllib3
